+++ CHANGELOG.md
@@ -9,10 +9,12 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 **New fixes in 1.45:**
 
 * [mse274424: Social login buttons CSS on the login/register page break when zoom less than 100%](http://meta.stackexchange.com/q/274424)
+* [mse275976: Unable to see completed badges](http://meta.stackexchange.com/q/275976) (credit: Patrick Hofman)
 
 **Other changes:**
 
 * Improved [mse259692](http://meta.stackexchange.com/q/259692) fix by hacking the JS Date constructor and methods to pretend that local time is UTC.
+
 
 1.44 (5 Feb 2016)
 ====